Examples: Rename lib & Add dev mode
Add dev mode to examples that make working on the easier.
Rename experiments library to examples.
