+++ src/faebryk/libs/examples/resources/example/example.kicad_pcb
